# sign and verify contracts with secp256k1