--- a/bench/index.js
+++ b/bench/index.js
@@ -116,6 +116,11 @@ function signMessage(deferred) {
   });
 }
 
+function signContract(deferred) {
+  contract1.sign('renter', renterPrivateKey);
+  deferred.resolve();
+}
+
 suite.add('new hd contract', newHDContract, {maxTime: maxTime});
 suite.add('new contract', newContract, {maxTime: maxTime});
 suite.add('verify message with hd contact', verifyHDContact, {
@@ -127,6 +132,10 @@ suite.add('verify message with contact', verifyContact, {
   defer: true
 });
 suite.add('sign message', signMessage, {
+  maxTime: maxTime,
+  defer: true
+});
+suite.add('sign contract', signContract, {
   maxTime: maxTime,
   defer: true
 });