--- a/bench/index.js
+++ b/bench/index.js
@@ -121,6 +121,11 @@ function signContract(deferred) {
   deferred.resolve();
 }
 
+function verifyContract(deferred) {
+  contract1.verify('renter', renterID);
+  deferred.resolve();
+}
+
 suite.add('new hd contract', newHDContract, {maxTime: maxTime});
 suite.add('new contract', newContract, {maxTime: maxTime});
 suite.add('verify message with hd contact', verifyHDContact, {
@@ -136,6 +141,10 @@ suite.add('sign message', signMessage, {
   defer: true
 });
 suite.add('sign contract', signContract, {
+  maxTime: maxTime,
+  defer: true
+});
+suite.add('verify contract', verifyContract, {
   maxTime: maxTime,
   defer: true
 });