--- a/lib/contract/index.js
+++ b/lib/contract/index.js
@@ -1,14 +1,15 @@
 'use strict';
 
-var assert = require('assert');
-var crypto = require('crypto');
-var merge = require('merge');
-var JSONSchema = require('jsen');
-var stringify = require('json-stable-stringify');
-var bitcore = require('bitcore-lib');
-var constants = require('../constants');
-var Message = require('bitcore-message');
-var ms = require('ms');
+const assert = require('assert');
+const crypto = require('crypto');
+const merge = require('merge');
+const JSONSchema = require('jsen');
+const stringify = require('json-stable-stringify');
+const bitcore = require('bitcore-lib');
+const constants = require('../constants');
+const secp256k1 = require('secp256k1');
+const Message = require('bitcore-message');
+const ms = require('ms');
 
 /**
  * Represents a storage contract between a renter and a farmer
@@ -258,8 +259,17 @@ Contract.prototype.verify = function(actor, pubkeyhash) {
  * @returns {String} externalSignature
  */
 Contract.prototype.signExternal = function(secret) {
-  var message = Message(this.getSigningData());
-  return message.sign(bitcore.PrivateKey.fromString(secret));
+  var keypair = bitcore.PrivateKey.fromString(secret);
+  var hash = Message(this.getSigningData()).magicHash();
+  var signobj = secp256k1.sign(hash,
+    keypair.toBuffer());
+  var sign = bitcore.crypto.Signature.fromDER(
+    secp256k1.signatureExport(signobj.signature)
+  ).toCompact(
+    signobj.recovery, 
+    keypair.toPublicKey().compressed
+  ).toString('base64');
+  return sign;
 };
 
 /**