--- a/lib/contract/index.js
+++ b/lib/contract/index.js
@@ -283,14 +283,19 @@ Contract.prototype.verifyExternal = function(signature, pubkeyhash) {
     return false;
   }
 
-  var message = Message(this.getSigningData());
-  var address = bitcore.Address.fromPublicKeyHash(Buffer(pubkeyhash, 'hex'));
-
+  var magic = Message(this.getSigningData()).magicHash();
+  
   try {
-    return message.verify(address, signature);
+    var signobj = bitcore.crypto.Signature.fromCompact(
+      new Buffer(signature, 'base64'));
+    var sig = secp256k1.signatureImport(signobj.toBuffer());
+    var recovery = signobj.i;
+    var pubKey = secp256k1.recover(magic, sig, recovery, true);
   } catch (err) {
     return false;
   }
+  return secp256k1.verify(magic, sig, pubKey) &&
+    bitcore.crypto.Hash.sha256ripemd160(pubKey).toString('hex') === pubkeyhash;
 };
 
 /**