--- a/lib/contract/index.js
+++ b/lib/contract/index.js
@@ -7,6 +7,7 @@ const JSONSchema = require('jsen');
 const stringify = require('json-stable-stringify');
 const bitcore = require('bitcore-lib');
 const constants = require('../constants');
+const KeyPair = require('../crypto-tools/keypair');
 const secp256k1 = require('secp256k1');
 const Message = require('bitcore-message');
 const ms = require('ms');
@@ -259,15 +260,15 @@ Contract.prototype.verify = function(actor, pubkeyhash) {
  * @returns {String} externalSignature
  */
 Contract.prototype.signExternal = function(secret) {
-  var keypair = bitcore.PrivateKey.fromString(secret);
+  var keypair = KeyPair(secret);
   var hash = Message(this.getSigningData()).magicHash();
   var signobj = secp256k1.sign(hash,
-    keypair.toBuffer());
+    new Buffer (keypair.getPrivateKeyPadded(), 'hex'));
   var sign = bitcore.crypto.Signature.fromDER(
     secp256k1.signatureExport(signobj.signature)
   ).toCompact(
     signobj.recovery, 
-    keypair.toPublicKey().compressed
+    keypair.getPublicKey.compressed
   ).toString('base64');
   return sign;
 };