For a better performance on the bridge side on sending ALLOC and renew contracts.

- [x] sign contract
- [X] benchmark test sign contracts
old: `sign contract x 16.83 ops/sec ±5.17% (80 runs sampled)`
new: `sign contract x 122 ops/sec ±1.54% (141 runs sampled)`
- [x] verify contracts
- [x] benachmark test verify contracts
old: `verify contract x 21.87 ops/sec ±1.33% (98 runs sampled)`
new: `verify contract x 632 ops/sec ±0.81% (147 runs sampled)`
- [x] unit test private key padding